+++ src/index.ts
@@ -2,5 +2,4 @@ import "reflect-metadata";
 
 export * from "./koa";
 export * from "./openapi";
-export * from "./mongoose";
 export * from "./references";