+++ src/openapi/index.ts
@@ -2,5 +2,5 @@ export { OpenApiResponse } from "./types";
 
 export * from "./decorators";
 export * from "./functions";
-export { IsDefinition } from "./definitions";
+export { ComponentSchema } from "./component-schema";
 export { OpenApi } from "./main";